fix(e2e): PR #532 round 4 — break NewsClientPort import cycle, normalize fake symbol, cast comment
